task(#13): Cleanup unused macro warnings
- Cannot be cleaned up per-se, but I can move the warning to only apply to C++
  code, where it may be more applicable anyway. There is no way to get around
  this warning in C for optimized builds when the debug framework I'm using is
  compiled out. I would need to switch to something else which would be a lot of
  work--perhaps in the future.

+++ include/rcsw/ds/bin_heap.h
@@ -139,7 +139,7 @@ static inline void* bin_heap_peek(const struct bin_heap * heap) {
 }
 
 static inline size_t bin_heap_height(const struct bin_heap * heap) {
-    return log10(bin_heap_n_elts(heap)) / log10(2);
+  return (size_t)(log10(bin_heap_n_elts(heap)) / log10(2));
 }
 
 /*******************************************************************************
